separate fob tests into different file
