+++ tensorboard/webapp/widgets/linked_time_fob/BUILD
@@ -40,6 +40,7 @@ tf_ts_library(
     testonly = True,
     srcs = [
         "linked_time_fob_test.ts",
+        "linked_time_fob_controller_test.ts",
     ],
     deps = [
         ":linked_time_fob",
